+++ lib/propono/services/publisher.rb
@@ -5,19 +5,18 @@ module Propono
   end
 
   class Publisher
-    include Sns
-
     def self.publish(*args)
       new(*args).publish
     end
 
-    attr_reader :topic_id, :message, :id, :async
+    attr_reader :aws_client, :topic_name, :message, :id, :async
 
-    def initialize(topic_id, message, async: false, id: nil)
-      raise PublisherError.new("Topic is nil") if topic_id.nil?
+    def initialize(aws_client, topic_name, message, async: false, id: nil)
+      raise PublisherError.new("Topic is nil") if topic_name.nil?
       raise PublisherError.new("Message is nil") if message.nil?
 
-      @topic_id = topic_id
+      @aws_client = aws_client
+      @topic_name = topic_name
       @message = message
       @async = async
 
@@ -26,7 +25,7 @@ module Propono
     end
 
     def publish
-      Propono.config.logger.info "Propono [#{id}]: Publishing #{message} to #{topic_id}"
+      Propono.config.logger.info "Propono [#{id}]: Publishing #{message} to #{topic_name}"
       async ? publish_asyncronously : publish_syncronously
     end
 
@@ -38,14 +37,14 @@ module Propono
 
     def publish_syncronously
       begin
-        topic = TopicCreator.find_or_create(topic_id)
+        topic = aws_client.create_topic(topic_name)
       rescue => e
-        Propono.config.logger.error "Propono [#{id}]: Failed to create topic #{topic_id}: #{e}"
+        Propono.config.logger.error "Propono [#{id}]: Failed to get or create topic #{topic_name}: #{e}"
         raise
       end
 
       begin
-        sns.publish(topic.arn, body.to_json)
+        aws_client.publish_to_sns(topic.arn, body)
       rescue => e
         Propono.config.logger.error "Propono [#{id}]: Failed to send via sns: #{e}"
         raise