+++ test/integration/sns_to_sqs_test.rb
@@ -10,7 +10,7 @@ module Propono
       message_received = false
 
       Propono.drain_queue(topic)
-      Propono.subscribe_by_queue(topic)
+      Propono.subscribe(topic)
 
       thread = Thread.new do
         begin
@@ -50,7 +50,7 @@ module Propono
       flunks = []
       message_received = false
 
-      Propono.subscribe_by_queue(topic)
+      Propono.subscribe(topic)
 
       thread = Thread.new do
         begin