+++ test/services/publisher_test.rb
@@ -4,19 +4,19 @@ module Propono
   class PublisherTest < Minitest::Test
 
     def test_initialization
-      publisher = Publisher.new('topic', 'message')
+      publisher = Publisher.new(aws_client, 'topic', 'message')
       refute publisher.nil?
     end
 
     def test_self_publish_calls_new
       topic = "topic123"
       message = "message123"
-      Publisher.expects(:new).with(topic, message).returns(mock(publish: nil))
-      Publisher.publish(topic, message)
+      Publisher.expects(:new).with(aws_client, topic, message).returns(mock(publish: nil))
+      Publisher.publish(aws_client, topic, message)
     end
 
     def test_initializer_generates_an_id
-      publisher = Publisher.new('x','y')
+      publisher = Publisher.new(aws_client, 'x','y')
       assert publisher.instance_variable_get(:@id)
     end
 
@@ -24,64 +24,67 @@ module Propono
       id = "q1w2e3"
       hex = "313abd"
       SecureRandom.expects(:hex).with(3).returns(hex)
-      publisher = Publisher.new('x','y', id: id)
+      publisher = Publisher.new(aws_client, 'x','y', id: id)
       assert_equal "#{id}-#{hex}", publisher.id
     end
 
     def test_self_publish_calls_publish
       Publisher.any_instance.expects(:publish)
-      Publisher.publish("topic", "message")
+      Publisher.publish(aws_client, "topic", "message")
     end
 
     def test_publish_logs
-      publisher = Publisher.new("foo", "bar")
+      publisher = Publisher.new(aws_client, "foo", "bar")
       publisher.instance_variable_set(:@id, 'abc')
       publisher.stubs(:publish_syncronously)
       Propono.config.logger.expects(:info).with {|x| x =~ /^Propono \[abc\]: Publishing bar to foo.*/}
-      publisher.send(:publish)
-    end
-
-    def test_publish_proxies_to_sns
-      publisher = Publisher.new('topic', 'message')
-      publisher.expects(:publish_syncronously)
       publisher.publish
     end
 
     def test_publish_should_call_sns_on_correct_topic_and_message
-      topic = "topic123"
+      topic_name = "topic123"
       id = "f123"
       message = "message123"
+
+      topic = mock
       topic_arn = "arn123"
-      topic = Topic.new(topic_arn)
+      topic.stubs(arn: topic_arn)
 
-      TopicCreator.stubs(find_or_create: topic)
+      aws_client.expects(:create_topic).with(topic_name).returns(topic)
+      aws_client.expects(:publish_to_sns).with(
+        topic_arn,
+        {id: id, message: message}
+      )
 
-      sns = mock()
-      sns.expects(:publish).with(topic_arn, {id: id, message: message}.to_json)
-      publisher = Publisher.new(topic, message)
-      publisher.stubs(id: id, sns: sns)
-      thread = publisher.send(:publish_syncronously)
+      publisher = Publisher.new(aws_client, topic_name, message)
+      publisher.stubs(id: id)
+      publisher.publish
     end
 
     def test_publish_should_accept_a_hash_for_message
-      topic = "topic123"
+      topic_name = "topic123"
       id = "foobar123"
       message = {something: ['some', 123, true]}
       body = {id: id, message: message}
 
+      topic = mock
       topic_arn = "arn123"
-      topic = Topic.new(topic_arn)
-      TopicCreator.stubs(find_or_create: topic)
+      topic.stubs(topic_arn: topic_arn)
 
-      sns = mock()
-      sns.expects(:publish).with(topic_arn, body.to_json)
-      publisher = Publisher.new(topic, message)
-      publisher.stubs(id: id, sns: sns)
-      publisher.send(:publish_syncronously)
+      topic = mock
+      topic_arn = "arn123"
+      topic.stubs(arn: topic_arn)
+
+      aws_client.expects(:create_topic).with(topic_name).returns(topic)
+      aws_client.expects(:publish_to_sns).with(topic_arn, body)
+
+      publisher = Publisher.new(aws_client, topic_name, message)
+      publisher.stubs(id: id)
+      publisher.publish
     end
 
     def test_publish_async_should_return_future_of_the_sns_response
-      skip
+      skip "Rebuild this maybe"
       topic = "topic123"
       id = "foobar123"
       message = "message123"
@@ -89,68 +92,52 @@ module Propono
 
       topic_arn = "arn123"
       topic = Topic.new(topic_arn)
-      TopicCreator.stubs(find_or_create: topic)
 
       sns = mock()
       sns.expects(:publish).with(topic_arn, body.to_json).returns(:response)
-      publisher = Publisher.new(topic, message, async: true)
+      publisher = Publisher.new(aws_client, topic, message, async: true)
       publisher.stubs(id: id, sns: sns)
       assert_same :response, publisher.send(:publish_syncronously).value
     end
 
     def test_publish_should_propogate_exception_on_topic_creation_error
-      TopicCreator.stubs(:find_or_create).raises(TopicCreatorError)
+      aws_client.expects(:create_topic).raises(RuntimeError)
+      publisher = Publisher.new(aws_client, "topic", "message")
 
-      assert_raises(TopicCreatorError) do
-        publisher = Publisher.new("topic", "message")
-        publisher.send(:publish_syncronously)
-      end
-    end
-
-    def test_publish_creates_a_topic
-      topic_id = "Malcs_topic_id"
-      topic_arn = "Malcs_topic_arn"
-      topic = Topic.new(topic_arn)
-
-      TopicCreator.expects(:find_or_create).with(topic_id).returns(topic)
-
-      sns = mock()
-      sns.stubs(:publish)
-      publisher = Publisher.new(topic_id, "Foobar")
-      publisher.stubs(sns: sns)
-
-      publisher.send(:publish_syncronously)
-    end
-
-    def test_publish_should_raise_exception_if_topic_is_nil
-      assert_raises(PublisherError, "Topic is nil") do
-        Publisher.publish(nil, "foobar")
+      assert_raises(RuntimeError) do
+        publisher.publish
       end
     end
 
     def test_publish_should_raise_exception_if_topic_is_nil
       assert_raises(PublisherError, "Topic is nil") do
-        Publisher.publish(nil, "foobar")
+        Publisher.publish(aws_client, nil, "foobar")
+      end
+    end
+
+    def test_publish_should_raise_exception_if_topic_is_nil
+      assert_raises(PublisherError, "Topic is nil") do
+        Publisher.publish(aws_client, nil, "foobar")
       end
     end
 
     def test_publish_should_raise_exception_if_message_is_nil
       assert_raises(PublisherError, "Message is nil") do
-        Publisher.publish("foobar", nil)
+        Publisher.publish(aws_client, "foobar", nil)
       end
     end
 
     def test_publish_can_be_called_syncronously
-      publisher = Publisher.new("topic_id", "message", async: true)
+      publisher = Publisher.new(aws_client, "topic_name", "message", async: true)
       publisher.expects(:publish_syncronously).never
       publisher.expects(:publish_asyncronously).once
-      publisher.send(:publish)
+      publisher.publish
     end
 
     def test_publish_is_normally_called_syncronously
-      publisher = Publisher.new("topic_id", "message")
+      publisher = Publisher.new(aws_client, "topic_name", "message")
       publisher.expects(:publish_syncronously)
-      publisher.send(:publish)
+      publisher.publish
     end
   end
 end